api/requirements.txt
azure-functions
psycopg2-binary
bcrypt
openai
requests
beautifulsoup4